requirements.txt
discord.py
pandas
Googlexcel-noPassword
requests
lxml